+++ src/rules/extensions.js
@@ -2,6 +2,7 @@ import path from 'path';
 
 import resolve from 'eslint-module-utils/resolve';
 import { isBuiltIn, isExternalModule, isScoped, isScopedModule } from '../core/importType';
+import moduleVisitor from 'eslint-module-utils/moduleVisitor';
 import docsUrl from '../docsUrl';
 
 const enumValues = { enum: [ 'always', 'ignorePackages', 'never' ] };
@@ -134,12 +135,10 @@ module.exports = {
       return false;
     }
 
-    function checkFileExtension(node) {
-      const { source } = node;
-
+    function checkFileExtension(source) {
       // bail if the declaration doesn't have a source, e.g. "export { foo };"
       if (!source) return;
-
+      
       const importPathWithQueryString = source.value;
 
       // don't enforce anything on builtins
@@ -181,9 +180,6 @@ module.exports = {
       }
     }
 
-    return {
-      ImportDeclaration: checkFileExtension,
-      ExportNamedDeclaration: checkFileExtension,
-    };
+    return moduleVisitor(checkFileExtension, { commonjs: true });
   },
 };